+++ src/app/actions.js
@@ -1,4 +1,4 @@
-import { CHANGE_SEARCH_FIELD, REQUEST_ROBOTS_PENDING, REQUEST_ROBOTS_SUCCESS, REQUEST_ROBOTS_FAILED } from "./constants";
+import { CHANGE_SEARCH_FIELD, REQUEST_ROBOTS_PENDING, REQUEST_ROBOTS_SUCCESS, REQUEST_ROBOTS_FAILED } from "../constants";
 
 export const setSearchField = (text) => ({
   type: CHANGE_SEARCH_FIELD,
