+++ src/index.js
@@ -3,9 +3,9 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { createStore, applyMiddleware, combineReducers } from "redux";
 import { thunk } from "redux-thunk";
-import { createLogger } from "redux-logger";
+import { store } from "./app/store";
 
-import { searchRobots, requestRobots } from "./reducers";
+import { searchRobots, requestRobots } from "./app/reducers";
 
 import "./index.css";
 import App from "./containers/App";
@@ -13,8 +13,7 @@ import reportWebVitals from "./reportWebVitals";
 import "tachyons";
 
 const rootReducer = combineReducers({ searchRobots, requestRobots });
-const logger = createLogger();
-const store = createStore(rootReducer, applyMiddleware(thunk, logger));
+// const store = createStore(rootReducer, applyMiddleware(thunk, logger));
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
